third_party/WebKit/LayoutTests/fast/canvas/canvas-font-cache.html

<script src="../../resources/js-test.js"></script>
<canvas id='c'></canvas>
<script>
// The reason this is implemented as a layout test instead of a unit test (see CanvasFontCacheTest.cpp)
// is to exercise interactions with animation frames.

var cacheLimit = 0;
if (!window.internals || !window.testRunner) {
	testFailed('This test requires the window.internals and window.testRunner interfaces');
} else {
	cacheLimit = internals.canvasFontCacheMaxFonts();
	testRunner.waitUntilDone();
	testRunner.dumpAsText();
	requestAnimationFrame(step1);
}

var ctx = document.getElementById('c').getContext('2d');

function step1()
{
	for (var i = 0; i < cacheLimit + 1; i ++) {
		ctx.font = (1 + i) + 'px sans-serif';
		ctx.fillText('a', 0, 50);
	}
	shouldBeTrue('internals.isInCanvasFontCache(document, "1px sans-serif")');
	requestAnimationFrame(step2);
}

function step2()
{
	// Pruning of the cache is expected to happen between tasks
	shouldBeFalse('internals.isInCanvasFontCache(document, "1px sans-serif")');
	shouldBeTrue('internals.isInCanvasFontCache(document, "2px sans-serif")'); 
	testRunner.notifyDone();
}

</script>
